// apps/backend/src/api/schema.ts
export interface components {
    schemas: {
        Availability: {
            slots: components["schemas"]["Slot"][];
        };
        Booking: {
            /** Format: uuid */
            id: string;
            /** Format: uuid */
            eventTypeId: string;
            /** Format: date-time */
            startAt: string;
            guestName: string;
            /** Format: email */
            guestEmail: string;
        };
        BookingCreate: {
            /** Format: uuid */
            eventTypeId: string;
            /** Format: date-time */
            startAt: string;
            guestName: string;
            /** Format: email */
            guestEmail: string;
        };
        BookingList: {
            bookings: components["schemas"]["BookingListItem"][];
        };
        BookingListItem: {
            eventType: components["schemas"]["EventTypeRef"];
            /** Format: date-time */
            startAt: string;
        };
        ErrorBody: {
            error: {
                /** @enum {string} */
                code: "VALIDATION_ERROR" | "INVALID_ADMIN_PASSWORD" | "EVENT_TYPE_NOT_FOUND" | "SLOT_TAKEN";
                message: string;
            };
        };
        EventType: {
            /** Format: uuid */
            id: string;
            name: string;
            description: string;
            /** Format: int32 */
            durationMinutes: number;
        };
        EventTypeCreate: {
            name: string;
            description: string;
            /** Format: int32 */
            durationMinutes: number;
        };
        EventTypeList: {
            eventTypes: components["schemas"]["EventType"][];
        };
        EventTypeRef: {
            /** Format: uuid */
            id: string;
            name: string;
        };
        Slot: {
            /** Format: date-time */
            startAt: string;
            available: boolean;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}